Extracting label styling into a new component ActivityCardLabels
